+++ src/k8s_sandbox/_helm.py
@@ -21,7 +21,11 @@ DEFAULT_CHART = Path(__file__).parent / "resources" / "helm" / "agent-env"
 DEFAULT_TIMEOUT = 300
 MAX_INSTALL_ATTEMPTS = 3
 INSTALL_RETRY_DELAY_SECONDS = 5
-
+INSPECT_HELM_TIMEOUT = "INSPECT_HELM_TIMEOUT"
+HELM_CONTEXT_DEADLINE_EXCEEDED_URL = (
+    "https://k8s-sandbox.ai-safety-institute.org.uk/tips/troubleshooting/"
+    "#helm-context-deadline-exceeded"
+)
 
 logger = logging.getLogger(__name__)
 
@@ -147,6 +151,16 @@ class Release:
                 error=result.stderr,
             )
             raise _ResourceQuotaModifiedError(result.stderr)
+        if re.search(r"INSTALLATION FAILED: context deadline exceeded", result.stderr):
+            _raise_runtime_error(
+                f"Helm install timed out (context deadline exceeded). The configured "
+                f"timeout value was {_get_timeout()}s. Please see the docs for why "
+                f"this might occur: {HELM_CONTEXT_DEADLINE_EXCEEDED_URL}. Also "
+                f"consider increasing the timeout by setting the "
+                f"{INSPECT_HELM_TIMEOUT} environment variable.",
+                release=self.release_name,
+                result=result,
+            )
         _raise_runtime_error(
             "Helm install failed.", release=self.release_name, result=result
         )
@@ -242,15 +256,10 @@ async def _run_subprocess(
 
 
 def _get_timeout() -> int:
-    if user_configured_timeout := os.environ.get("INSPECT_HELM_TIMEOUT"):
-        timeout_int = int(user_configured_timeout)
-        if timeout_int <= 0:
-            raise ValueError(
-                "INSPECT_HELM_TIMEOUT must be a positive int: "
-                f"{user_configured_timeout}"
-            )
-        return timeout_int
-    return DEFAULT_TIMEOUT
+    timeout = _get_environ_int(INSPECT_HELM_TIMEOUT, DEFAULT_TIMEOUT)
+    if timeout <= 0:
+        raise ValueError(f"{INSPECT_HELM_TIMEOUT} must be a positive int: '{timeout}'.")
+    return timeout
 
 
 def _install_semaphore() -> asyncio.Semaphore:
@@ -272,5 +281,7 @@ def _uninstall_semaphore() -> asyncio.Semaphore:
 def _get_environ_int(name: str, default: int) -> int:
     try:
         return int(os.environ[name])
-    except (KeyError, ValueError):
+    except KeyError:
         return default
+    except ValueError as e:
+        raise ValueError(f"{name} must be an int: '{os.environ[name]}'.") from e